Before kstring

+++ libc/syscalls/Linux/tasking/exec.cpp
@@ -30,7 +30,7 @@ SOFTWARE.
 
 extern int common_syscall(size_t type, size_t no, ...);
 
-long execve(const char* path, const char* argv[], const char* envp[])
+int execve(const char* path, const char* argv[], const char* envp[])
 {
     auto ret = common_syscall(1, SYS_execve, path, argv, envp);
     if (ret < 0)
